refactor: Moved create poll action
