+++ src/unipoll_api/actions/workspace.py
@@ -1,12 +1,9 @@
-# from typing import Optional
-# from pydantic import EmailStr
-from beanie import WriteRules
 from unipoll_api import AccountManager
 from unipoll_api import actions
-from unipoll_api.documents import Workspace, Account, Policy, Poll
+from unipoll_api.documents import Workspace, Account, Policy
 from unipoll_api.utils import Permissions
-from unipoll_api.schemas import WorkspaceSchemas, PollSchemas
-from unipoll_api.exceptions import (WorkspaceExceptions, PollExceptions)
+from unipoll_api.schemas import WorkspaceSchemas
+from unipoll_api.exceptions import WorkspaceExceptions
 
 
 # Get a list of workspaces where the account is a owner/member
@@ -38,8 +35,7 @@ async def create_workspace(input_data: WorkspaceSchemas.WorkspaceCreateInput) ->
     if not new_workspace:
         raise WorkspaceExceptions.ErrorWhileCreating(input_data.name)
 
-    await new_workspace.add_member(account=account, permissions=Permissions.WORKSPACE_ALL_PERMISSIONS, save=False)
-    await Workspace.save(new_workspace, link_rule=WriteRules.WRITE)
+    await new_workspace.add_member(account=account, permissions=Permissions.WORKSPACE_ALL_PERMISSIONS)
 
     # Specify fields for output schema
     return WorkspaceSchemas.WorkspaceCreateOutput(**new_workspace.model_dump(include={'id', 'name', 'description'}))
@@ -54,7 +50,7 @@ async def get_workspace(workspace: Workspace,
     groups = (await actions.GroupActions.get_groups(workspace)).groups if include_groups else None
     members = (await actions.MembersActions.get_members(workspace)).members if include_members else None
     policies = (await actions.PolicyActions.get_policies(resource=workspace)).policies if include_policies else None
-    polls = (await get_polls(workspace)).polls if include_polls else None
+    polls = (await actions.PollActions.get_polls(workspace)).polls if include_polls else None
     # Return the workspace with the fetched resources
     return WorkspaceSchemas.Workspace(id=workspace.id,
                                       name=workspace.name,
@@ -97,37 +93,3 @@ async def delete_workspace(workspace: Workspace):
     if await workspace.get(workspace.id):
         raise WorkspaceExceptions.ErrorWhileDeleting(workspace.id)
     await Policy.find({"parent_resource._id": workspace.id}, fetch_links=True).delete()
-
-
-# Get a list of polls in a workspace
-async def get_polls(workspace: Workspace) -> PollSchemas.PollList:
-    return await actions.PollActions.get_polls(workspace)
-
-
-# Create a new poll in a workspace
-async def create_poll(workspace: Workspace, input_data: PollSchemas.CreatePollRequest) -> PollSchemas.PollResponse:
-    # Check if poll name is unique
-    poll: Poll  # For type hinting, until Link type is supported
-    for poll in workspace.polls:  # type: ignore
-        if poll.name == input_data.name:
-            raise PollExceptions.NonUniqueName(poll)
-
-    # Create a new poll
-    new_poll = Poll(name=input_data.name,
-                    description=input_data.description,
-                    workspace=workspace,  # type: ignore
-                    public=input_data.public,
-                    published=input_data.published,
-                    questions=input_data.questions,
-                    policies=[])
-
-    # Check if poll was created
-    if not new_poll:
-        raise PollExceptions.ErrorWhileCreating(new_poll)
-
-    # Add the poll to the workspace
-    workspace.polls.append(new_poll)  # type: ignore
-    await Workspace.save(workspace, link_rule=WriteRules.WRITE)
-
-    # Return the new poll
-    return PollSchemas.PollResponse(**new_poll.model_dump())
